backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication
python-telegram-bot>=21.0,<22.0

# Media & Files
Pillow>=10.2,<11.0

# Environment
python-dotenv>=1.0,<2.0

# Production
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0

# Development
django-debug-toolbar>=4.2,<5.0
